esp/espnow: Reimplement RSSI support.
Streamline the RSSI support. Do not update the device-table from
recv_cb(). Instead, save the rssi and timestamp in the packet buffer
and update the device-table when the packet is read from
irecv()/recv(). Also simplifies the device-table handling code.
Docs updated to reflect the changes:
 - RSSI values of first messages from peer are no longer lost
 - Device-table is only updated when irecv()/recv() read the message
    from the buffer.
Also for IDF < 4.2.0, rssi needs to be offset by 100.